refactor(electron): modularize main.js into focused modules (closes #749)
Split the 1,425-line electron/main.js monolith into 8 focused modules:
- app-constants.js: isDev, APP_NAME, isMicrosoftStoreApp
- window-manager.js: createWindow, allWindows tracking, power broadcast
- menu.js: buildApplicationMenu, rebuildApplicationMenu, menuUiState
- auto-updater.js: setupAutoUpdater, checkForUpdates
- quick-look.js: installQuickLookPluginIfNeeded
- ipc/file-ipc.js: open-file, save-file, export-*, get-pending-file, path security
- ipc/shell-ipc.js: file-manager, open-external, dictionary popup, context menu
- ipc/system-ipc.js: chrome-version, dirty, safe-storage, power, menu IPC

main.js is now a 154-line entry point handling only ASAR setup,
single-instance lock, open-file event, and app lifecycle.
Circular deps between window-manager/menu/auto-updater resolved via
deferred require() inside function bodies.
